Add config for home URL that browser is redirected
What
===
Add config for home URL that the client is redirected to if the client
is a browser.

Why
===
As it stands vangen redirects the browser to the godocs for the package.
This makes sense for libraries but not so much for CLIs. In those cases
it might make sense to redirect to a website for the product.